requirements.txt
langchain
langchain-google-genai
langchain-community
faiss-cpu
pypdf
streamlit
python-dotenv
google-generativeai